+++ src/sbs/trigger.cpp
@@ -252,7 +252,7 @@ bool Trigger::DoAction()
 	std::vector<Action*> actionlist = sbs->GetAction(Actions[current_position - 1]);
 
 	bool result = true;
-	for (int i = 0; i < actionlist.size(); i++)
+	for (int i = 0; i < (int)actionlist.size(); i++)
 	{
 		bool result2 = false;
